+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * Core: Add HTTP and TCP Services and require them running
+* Core: Add loading environment variables from the system
 
 <!--
 ### Added